Refactor functions to new file; #1088
